% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcToyExampleVisualisation.R
\name{PlotEpsilonTrace}
\alias{PlotEpsilonTrace}
\title{Plot epsilon trace.}
\usage{
PlotEpsilonTrace(smc.result, state.to.visualise, use.run.length, use.eps.max)
}
\description{
Plot epsilon trace.
}

