% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcToyExampleVisualisation.R
\name{PlotHistogram}
\alias{PlotHistogram}
\title{Plot histogram.}
\usage{
PlotHistogram(thetas, use.thetas)
}
\description{
Note that the thetas are assumed to have been resampled
before being used as input here.
}

