% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maExample.R
\name{smcMovingAverageExample}
\alias{smcMovingAverageExample}
\title{MA(2) example.}
\usage{
smcMovingAverageExample(use.raw.distance.function = T,
  create.debug.variables = F, debug.space = 1)
}
\description{
MA(2) example.
}

